Minor fix (added notification) on updater.

+++ src/utils/update.py
@@ -30,35 +30,39 @@ from src.thirdparty.colorama import Fore, Back, Style, init
 def updater():
 
   time.sleep(1)
-  sys.stdout.write("(*) Checking requirements to update "+ settings.APPLICATION + " via Github... ")
+  sys.stdout.write("(*) Checking requirements to update "+ settings.APPLICATION + " via GitHub... ")
   sys.stdout.flush()
-  try:
-    # Check if git is installed
-    requirment = "git"
-    requirments.do_check(requirment)
-    if requirments.do_check(requirment) == True :
-      # Check if ".git" exists!
-      if os.path.isdir("./.git"):
-        # Check if git is installed
-        sys.stdout.write("[" + Fore.GREEN + " SUCCEED " + Style.RESET_ALL + "]\n")
-        sys.stdout.flush()
-        print "\n---"
-        subprocess.Popen("git reset --hard HEAD && git pull", shell=True).wait()
-        # Delete *.pyc files.
-        subprocess.Popen("find . -name \"*.pyc\" -delete", shell=True).wait()
-        # Delete empty directories and files.
-        subprocess.Popen("find . -empty -type d -delete", shell=True).wait()
-        print "---\n"
+  # Check if windows
+  if settings.IS_WINDOWS:
+    print "["+ Fore.RED + " FAILED " + Style.RESET_ALL +"]"
+    print Back.RED + "(x) Error: For updating purposes on Windows platform, it's recommended to use a GitHub client for Windows (http://windows.github.com/)." + Style.RESET_ALL
+    sys.exit(0)
+  else:
+    try:
+      requirment = "git"
+      # Check if 'git' is installed.
+      requirments.do_check(requirment)
+      if requirments.do_check(requirment) == True :
+        # Check if ".git" exists!
+        if os.path.isdir("./.git"):
+          sys.stdout.write("[" + Fore.GREEN + " SUCCEED " + Style.RESET_ALL + "]\n")
+          sys.stdout.flush()
+          print "\n---"
+          subprocess.Popen("git reset --hard HEAD && git pull", shell=True).wait()
+          # Delete *.pyc files.
+          subprocess.Popen("find . -name \"*.pyc\" -delete", shell=True).wait()
+          # Delete empty directories and files.
+          subprocess.Popen("find . -empty -type d -delete", shell=True).wait()
+          print "---\n"
+        else:
+          print "["+ Fore.RED + " FAILED " + Style.RESET_ALL +"]"
+          print Back.RED + "(x) Error: The '.git' directory not found. Do it manually: "+ Style.BRIGHT +"'git clone " + settings.GIT_URL + " " + settings.APPLICATION +"' "+ Style.RESET_ALL    
+          sys.exit(0)
       else:
-        print "["+ Fore.RED + " FAILED " + Style.RESET_ALL +"]"
+          print "["+ Fore.RED + " FAILED " + Style.RESET_ALL +"]"
+          print Back.RED + "(x) Error: " + requirment + " not found." + Style.RESET_ALL + "\n"
+          sys.exit(0)
 
-        print Back.RED + "(x) Error: The '.git' directory not found. Do it manually: "+ Style.BRIGHT +"'git clone " + settings.GIT_URL + " " + settings.APPLICATION +"' "+ Style.RESET_ALL    
-        sys.exit(0)
-    else:
-        print "["+ Fore.RED + " FAILED " + Style.RESET_ALL +"]"
-        print Back.RED + "(x) Error: " + requirment + " not found." + Style.RESET_ALL + "\n"
-        sys.exit(0)
-
-  except Exception as error:
-    print Back.RED + "\n(x) Error: " +str(error) + Style.RESET_ALL + "\n" 
-  sys.exit(0)
+    except Exception as error:
+      print Back.RED + "\n(x) Error: " + str(error) + Style.RESET_ALL + "\n" 
+    sys.exit(0)
